requirements.txt
# NEXUS GraphRAG - Requirements
# Python 3.11+ recommended
# Last Updated: December 11, 2025

# ===========================================
# CORE DATABASE CONNECTORS
# ===========================================
neo4j==5.14.0
pymilvus==2.3.4
snowflake-connector-python==3.0.4
databricks-sql-connector==2.9.0

# ===========================================
# LLM & LANGCHAIN (Compatible versions)
# ===========================================
# Using 0.2.x series for stability with langchain-ollama
langchain>=0.2.14,<0.3.0
langchain-ollama>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.20,<0.3.0

# ===========================================
# ML & EMBEDDINGS
# ===========================================
sentence-transformers==2.2.2
xgboost==2.0.3
scikit-learn==1.3.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0

# ===========================================
# DATA PROCESSING
# ===========================================
pandas>=2.0.0
pyarrow>=12.0.0

# ===========================================
# WEB INTERFACE
# ===========================================
gradio==4.10.0

# ===========================================
# UTILITIES
# ===========================================
python-dotenv==1.0.0
requests>=2.28.0
tqdm>=4.65.0
pydantic>=2.0.0

# ===========================================
# ASYNC & NETWORKING (for Milvus/Neo4j)
# ===========================================
aiohttp>=3.8.0
grpcio>=1.53.0
protobuf>=4.21.0

# ===========================================
# NOTES
# ===========================================
# 1. Run: pip install -r requirements.txt
# 2. For Apple Silicon (M1/M2), some packages may need:
#    pip install --no-cache-dir <package>
# 3. Milvus requires Docker: docker-compose up -d
# 4. Ollama must be installed separately: https://ollama.ai
#    Then run: ollama pull llama3.1